Split the unwritten-constants tests out of the transform module
The module crossed the 1000-line limit with its tests inline. The
qualifier lookup the module tests and the conformance tests both needed
moves into test_helpers so neither re-walks the library itself.
